modules were optimized (getRandomElement and getRandomSubArray was moved to utils.js)
