+++ index.html
@@ -12,8 +12,8 @@
 
   <main>
     <div class="promo">
-      <h1 class="promo__title visually-hidden">Keksobooking. Кексы по соседству</h1>
-      <img src="img/keksobooking.svg" alt="Keksobooking. Кексы по соседству" width="215" height="45">
+      <h1 class="promo__title">Keksobooking. Nurik Кексы по соседству</h1>
+      <img class="visually-hidden" src="img/keksobooking.svg" alt="Keksobooking. Кексы по соседству" width="215" height="45">
     </div>
 
     <!-- Карта объявлений -->
