+++ js/main.js
@@ -1,6 +1,7 @@
-import { getRandomAdvert } from './utils/get-random-advert.js';
+import { getRandomAdvert } from './data.js';
 
 const RANDOM_ADVERT_COUNT = 10;
 
 const randomAdverts = new Array(RANDOM_ADVERT_COUNT).fill(null).map(getRandomAdvert);
-randomAdverts;
+
+randomAdverts.forEach((elem)=> console.log(elem));
